Single and multi typeahead example based on Bootstrap CSS.
UX is similar to react-select.com. Halogen-HTML-CSS module is currently copied into this project as  it is not published for Halogen 5. When the browser loads the new example, the bulma.css file is dynamically removed from the page. If one of the old examples is loaded afterwards, that css file is missing, i.e. you should reload the page.